using PANDAS to extract and manipulate data
The file births.txt contains the number of births in Estonia for a certain time period by months, while the file deaths.txt contains the number of deaths. Write a program that finds and outputs the months when the natural population growth was positive and its size. Update your program so that it also finds and outputs the natural population growth for the entire period covered in the files.